// utils/types/types.ts
export type AffiliateUserData = {
    [x: string]: any;
    bizid: number;
    ztime: string;
    xcontact: string;
    xorg: string;
    xbranch: string;
    xphone: string;
    xmobile: string;
    xemail: string;
    xaddress1: string;
    xaddress2: string;
    xcountry: string;
    xstate: string;
    xcity: string;
    xsmsapiurl: string;
    xpromoemail: string;
    xsmtphost: string;
    xport: number;
    xsmtpuser: string;
    xsmtppass: string;
    xdomain: string;
    xdomainuser: string;
    xdomainpass: string;
    xsubdomain: string;
    xfbpageurl: string;
}